move theory layer to midigen/theory/
note, key, scale, roman, time_utils now live in
midigen/theory/. All imports updated across source
and tests.
