+++ midigen/__init__.py
@@ -1,14 +1,18 @@
-from .midigen import MidiGen
-from .track import Track
-from .note import Note
+from .theory.note import Note
+from .theory.key import Key, KEY_MAP
+from .theory.scale import Scale
+from .theory.time_utils import TimeConverter
+from .theory.roman import parse_roman_numeral, get_chord_pitches
+
 from .chord import Chord, ChordProgression, Arpeggio, ArpeggioPattern
-from .key import Key, KEY_MAP
-from .scale import Scale
 from .drums import DrumKit, Drum
-from .song import Song
-from .section import Section
-from .instruments import INSTRUMENT_MAP
-from .time_utils import TimeConverter
 from .melody import Melody
+from .section import Section
+
+from .track import Track
 from .channel_pool import ChannelPool, ChannelExhaustedError
+from .instruments import INSTRUMENT_MAP
+
+from .midigen import MidiGen
+from .song import Song
 from .compiler import MidiCompiler
