+++ midigen/midigen.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 import os
 from mido import MidiFile
-from midigen.key import Key
+from midigen.theory.key import Key
 from midigen.track import Track
 from pathlib import Path
