+++ midigen/theory/note.py
