+++ midigen/theory/scale.py
