+++ pyproject.toml
@@ -1,24 +1,23 @@
-[tool.poetry]
-name = "midigen-lib"   # distribution name on PyPI
-packages = [{ include = "midigen" }]
+[project]
+name = "midigen-lib"
 version = "1.0.0"
 description = "High-level, theory-powered MIDI generation library"
-authors = ["cainky <kylecain.me@gmail.com>"]
+authors = [{ name = "cainky", email = "kylecain.me@gmail.com" }]
 readme = "README.md"
 license = "GPL-3.0-only"
-repository = "https://github.com/cainky/midigen"
+requires-python = ">=3.10"
 keywords = ["midi", "music-generation", "mido", "music-theory"]
+dependencies = ["mido>=1.3.2,<2"]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-mido = "^1.3.2"
+[project.urls]
+Repository = "https://github.com/cainky/midigen"
 
-
-[tool.poetry.group.dev.dependencies]
-pre-commit = ">=3.6,<5.0"
-flake8 = ">=6.1,<8.0"
-pytest = "^9.0.1"
+[dependency-groups]
+dev = ["pre-commit>=3.6,<5.0", "flake8>=6.1,<8.0", "pytest>=9.0.1,<10"]
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["midigen"]
